Extract MockDependencyGraph to separate file
